requirements.txt
streamlit>=1.36
gspread
google-auth
